Revert ModManager
